feat: move trio and socket backends in a common backends sub-package.
